requirements.txt
Flask==2.3.2
python-dotenv==1.0.1
requests==2.31.0
python-telegram-bot==20.4
APScheduler==3.12.0
reportlab==4.0.0
SQLAlchemy==2.0.20